python/requirements.txt
easydict==1.9
numpy==1.17.0
tqdm==4.31.1
torchvision==0.5.0
torch==1.4.0
opencv_python==4.2.0.34
Pillow==7.1.2
tensorboardX==2.0
onnxruntime
attrdict
pandas
scipy
scikit-image
